# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Websters API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/v1/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authorization",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/positions/{id}": {
            "put": {
                "tags": [
                    "Positions"
                ],
                "summary": "Create a positions",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name_ua",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name_ru",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name_en",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Positions"
                ],
                "summary": "Delete a positions",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/positions": {
            "get": {
                "tags": [
                    "Positions"
                ],
                "summary": "Create a positions",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/positions/create": {
            "post": {
                "tags": [
                    "Positions"
                ],
                "summary": "Create a position",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name_ua",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name_ru",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name_en",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/users/{id} ": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/api/v1/users/create": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "{Auth}",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        }
    },
    "definitions": {}
}